checkout

+++ lib/controller/order.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 
 import 'package:http/http.dart' as http;
 
-generateTokenMidtrans(String orderId, int totalPrice) async {
+generateTokenMidtrans(String orderId, double totalPrice) async {
   String serverKey = "SB-Mid-server-5UKRWMkpzEb1Ds9ERKd6uo7Z";
 
   String url = 'https://app.sandbox.midtrans.com/snap/v1/transactions';
